move chart components to proper folder
