+++ src/lib/chart/PieChart.svelte
@@ -2,7 +2,7 @@
   import {onMount} from 'svelte';
   import Chart from 'chart.js/auto/auto.js';
   import {_} from 'svelte-i18n';
-  import type { ChartResult } from './customType';
+  import type { ChartResult } from '.../customType';
 
   export let ratioScope:ChartResult;
   let myChart;
